refactor: changed map store and marker service to handle the new config service files
